+++ summoner.py
@@ -34,7 +34,21 @@ class Summoner:
                 self.class_name.ljust(10, ' ') + \
                 "\t level " + str(self.level).rjust(2, ' ') + \
                 "\t" + xp_str.rjust(10, ' ') + " xp" + \
-                "\t" + str(round(self.gold)).rjust(6, ' ') + " gold"
+                "\t" + str(round(self.gold)).rjust(6, ' ') + " gold" + \
+                "\t" + "Next Adv: " + self.seconds_to_hms(self.time_to_next_adventure()) + \
+                "\t" + "Next Cellar: " + self.seconds_to_hms(self.time_to_next_cellar()) + \
+                " (" + str(self.expected_cellar_loot()) + " loot)"
+                
+
+    def seconds_to_hms(self, secs):
+        if (secs <= 0):
+            return "READY".ljust(9, ' ')
+        h = secs // 3600
+        secs = secs % 3600
+        m = secs // 60
+        s = secs % 60
+        hms = str(h) + "h" + str(m) + "m" + str(s) + "s"
+        return hms.ljust(9, ' ')
     
     def update_summoner_info(self):
         """Update class, level and xp"""
@@ -60,10 +74,14 @@ class Summoner:
         summoner_class = self.contracts["summoner"].functions.classes(index).call()
         return summoner_class
 
-    def check_adventure(self):    
+    def time_to_next_adventure(self):
+        """Get time to next adventure in seconds"""
         next_time_available = self.contracts["summoner"].functions.adventurers_log(self.token_id).call()
         current_time = self.transacter.timestamp
-        return current_time > next_time_available
+        return next_time_available - current_time
+
+    def check_adventure(self):    
+        return self.time_to_next_adventure() <= 0
 
     def force_adventure(self):
         print(Fore.WHITE + str(self) + " has gone on an adventure!")
@@ -117,14 +135,16 @@ class Summoner:
         if self.check_claim_gold():
             self.force_claim_gold()
 
-    def check_go_cellar(self):
+    def time_to_next_cellar(self):
         next_time_available = self.contracts["cellar"].functions.adventurers_log(self.token_id).call()
         current_time = self.transacter.timestamp
-        if current_time > next_time_available:
-            expected_loot = int(self.contracts["cellar"].functions.scout(self.token_id).call())
-            return expected_loot > 0
-        else:
-          return False
+        return next_time_available - current_time
+
+    def expected_cellar_loot(self):
+        return int(self.contracts["cellar"].functions.scout(self.token_id).call())
+
+    def check_go_cellar(self):
+        return self.time_to_next_cellar() <= 0 and self.expected_cellar_loot() > 0
 
     def force_go_cellar(self):
         print(Fore.WHITE + str(self) + " is going to The Cellar")